requirements.txt
torch>=1.13.0
torchvision>=0.14.0
diffusers>=0.11.1
transformers>=4.25.1
accelerate>=0.15.0
numpy>=1.23.5
Pillow>=9.3.0
matplotlib>=3.6.2
tqdm>=4.64.1
opencv-python>=4.6.0
ftfy>=6.1.1
scipy>=1.9.3 